refactor(ai): extract cache config from model cache
Moved CacheConfig, constants (DEFAULT_MODEL_REPO/FILE/SHA256), and
default_cache_dir() to cache_config.rs (137 lines).
model_cache.rs: 649 → 254 lines (-61%).
Re-exports maintained via ai/mod.rs for backward compat.
Updated semantic_cleaner_impl.rs and tokenizer.rs imports.
Both cargo check and cargo check --features ai pass clean.
